requirements.txt
beautifulsoup4
biopython
datasets
fake-useragent
flask
flask-cors
google-cloud-translate
joblib
lxml_html_clean
mysql-connector-python
news-paper3k
numpy
nltk
openpyxl
pandas
requests
scikit-learn
sentencepiece
torch
torchaudio
transformers
torchvision
gunicorn
tqdm